pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bluemark-fpv"
version = "0.1.0"
description = "BlueMark FPV — receiver-side identification aid for friendly FPV drone video. Hackathon demo. Identification aid only; human decision required."
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
]
ml = [
  "onnxruntime>=1.17",
  "ultralytics>=8.2",
  "pillow>=10",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
